+++ src/Model.jl
@@ -183,9 +183,9 @@ function prepare_shared_data(
     # All of these are expensive and independent of infection parameters
     init_state = initial_state(cd)
 
-    # add Rt compartment at end; this holds zero as r0 may differ across runs
+    # add new_vax tracking and Rt compartment at end
     init_state = reshape(init_state, length(init_state))
-    init_state = [init_state; 0.0]
+    init_state = [init_state; zeros(N_TOTAL_GROUPS); 0.0]
 
     contacts_unscaled = total_contacts(prepare_contacts(cd; scaled = false))
     cw = worker_contacts(cd)